refactors, layout changes, adding neccessary todo
* remove a console.log
* switch button layout to include the alternative way with keyboard
  shortcuts
* include a todo that marks an issue where the solution might depend on
  feedback
